fixed normal pertubations

+++ optimizers/mezo_torch.py
@@ -125,7 +125,7 @@ class MeZOOptimizer(Optimizer):
                 torch.manual_seed(self.zo_random_seed)
                 gen = torch.Generator(device= p.device)
                 z = torch.empty(p.shape).to(p.device)
-                z.normal_(mean=0, std=1, generator=gen, dtype=p.dtype)
+                z.normal_(mean=0, std=1, generator=gen).to(p.dtype)
                 print(p.device, p.dtype)
                 print(z.device, z.dtype)
                 p.add_(scaling_factor * zo_eps * z)
